+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Preflight connection check before opening website 
+
+### Changed
+
+- Removed Banner in CLI
+
 ## [1.0.5] - 2026-03-15
 
 ### Added